+++ bioprov/programs/__init__.py
@@ -2,7 +2,7 @@ __author__ = "Vini Salazar"
 __license__ = "MIT"
 __maintainer__ = "Vini Salazar"
 __url__ = "https://github.com/vinisalazar/bioprov"
-__version__ = "0.1.20"
+__version__ = "0.1.21"
 
 
 from .programs import (
@@ -12,6 +12,7 @@ from .programs import (
     blastp,
     muscle,
     mafft,
+    fasttree,
     prokka,
     kallisto_quant,
     kaiju,