+++ bioprov/workflows/__init__.py
@@ -2,10 +2,12 @@ __author__ = "Vini Salazar"
 __license__ = "MIT"
 __maintainer__ = "Vini Salazar"
 __url__ = "https://github.com/vinisalazar/bioprov"
-__version__ = "0.1.20"
+__version__ = "0.1.21"
 
 
-from .blastn import blastn_alignment
-from .genome_annotation import genome_annotation
-from .kaiju import KaijuWorkflow
-from .wf_parser import WorkflowOptionsParser
+from .workflows import (
+    blastn_alignment,
+    genome_annotation,
+    KaijuWorkflow,
+    WorkflowOptionsParser,
+)