+++ CHANGELOG.md
@@ -8,6 +8,15 @@
     * .write_paths_to_file, .copy_files_to_dir(), .link_files_to_dir() [ ]
 * Add logger calls when saving to JSON and uploading to ProvStore [ ]
 
+### v0.1.21
+* Add _config argument to `bp.load_project()` (this is a temporary fix) [x]
+* Add db property to Config class (prevents bug when setting DB path) [x]
+* Improve a few docstrings [x]
+* Refactor sha1 as sha256 [x]  
+* Add add and radd dunder methods for Directory and File [x]  
+* Package workflows in single module [x]
+* Add FastTree PresetProgram [x]
+
 ### v0.1.20
 * Debug graphical DOT output [x]
 * Add Muscle PresetProgram [x]